Nodes and connections can be copied and pasted -> next : fix the bug when a node has more than one input : the connection is not copied

+++ buttleofx/MainWindow.qml
@@ -32,10 +32,10 @@ ApplicationWindow {
     Keys.onPressed: {
 
         // Graph toolbar
-/*        if (event.key == Qt.Key_Delete) {
+        if (event.key == Qt.Key_Delete) {
            _buttleManager.deleteSelection();
         }
-*/        if ((event.key == Qt.Key_Z) && (event.modifiers & Qt.ControlModifier)) {
+       if ((event.key == Qt.Key_Z) && (event.modifiers & Qt.ControlModifier)) {
             if(_buttleManager.canUndo) {
                 _buttleManager.undo();
             }
@@ -224,6 +224,7 @@ ApplicationWindow {
                 onTriggered:
                     if(!_buttleData.currentSelectedNodeWrappers.isEmpty()) {
                         _buttleManager.nodeManager.copyNode()
+                        _buttleManager.connectionManager.copyConnections()
                     }
             }
 
@@ -233,6 +234,7 @@ ApplicationWindow {
                 onTriggered:
                     if(_buttleData.canPaste) {
                         _buttleManager.nodeManager.pasteNode();
+                        _buttleManager.connectionManager.pasteConnection()
                     }
             }
 
@@ -256,7 +258,7 @@ ApplicationWindow {
 
             MenuItem {
                 text: "Delete"
-                //shortcut: "del"
+                shortcut: "del"
                 onTriggered: _buttleManager.deleteSelection()
             }
         }
